Refactorización de la arquitectura del servidor
