+++ server/package.json
@@ -24,7 +24,9 @@
     "@nestjs/config": "3.3.0",
     "@nestjs/core": "^10.0.0",
     "@nestjs/platform-express": "^10.0.0",
+    "bcrypt": "^5.1.1",
     "cookie-parser": "^1.4.7",
+    "jsonwebtoken": "^9.0.2",
     "mysql2": "3.11.4",
     "reflect-metadata": "^0.2.0",
     "rxjs": "^7.8.1"
@@ -33,6 +35,7 @@
     "@nestjs/cli": "^10.0.0",
     "@nestjs/schematics": "^10.0.0",
     "@nestjs/testing": "^10.0.0",
+    "@types/bcrypt": "^5.0.2",
     "@types/cookie-parser": "^1.4.8",
     "@types/express": "^5.0.0",
     "@types/jest": "^29.5.2",
