+++ server/src/authentication/domain/constant/office.constants.ts
@@ -1,0 +1,8 @@
+export const Office = [
+  'Lima',
+  'Ica',
+  'Pisco',
+  'Chiclayo',
+  'Piura',
+  'Cusco'
+] as const;
