+++ server/src/authentication/domain/constant/role.constants.ts
@@ -1,0 +1,4 @@
+export const Role = [
+  'Administrador',
+  'Usuario'
+] as const;
