+++ server/src/authentication/domain/enums/default-user-value.enum.ts
@@ -1,0 +1,6 @@
+export enum DefaultUserValue {
+  Nickname = 'Nickname',
+  Password = 'Password',
+  Id = 0,
+  Dni = '99990000'
+}
