+++ server/src/authentication/domain/repository/user.repository.ts
@@ -1,0 +1,13 @@
+import UUID from "src/core/domain/value-object/uuid.value-object";
+import User from "../model/user.model";
+import Name from "src/core/domain/value-object/name.value-object";
+import Password from "../value-object/password.value-object";
+
+export const USER_REPOSITORY = Symbol('USER_REPOSITORY');
+
+export default interface IUserRepository {
+  findUserById(id: UUID): Promise<User>;
+  findUserIdByAuthentication(username: Name, password: Password): Promise<UUID>;
+  saveUser(user: User): Promise<void>;
+  deleteUser(id: UUID): Promise<void>;
+}
