+++ server/src/authentication/domain/value-object/password.value-object.ts
