+++ server/src/authentication/infrastructure/web/request/user.request.ts
@@ -1,0 +1,5 @@
+import { Request } from "express";
+
+export interface UserRequest extends Request {
+  user_id: string;
+}
