+++ server/src/core/domain/value-object/doi.value-object.ts
