+++ server/src/core/domain/value-object/email.value-object.ts
