+++ server/src/core/infrastructure/database/migration/implementation/costumer-type.migration.ts
