+++ server/src/core/infrastructure/database/migration/implementation/sale.migration.ts
