+++ server/src/core/infrastructure/error/mysql-connection.error.ts
@@ -1,0 +1,5 @@
+export default class MysqlConnectionError extends Error {
+  constructor(msg: string) {
+    super('Error de Conexion: ' + msg);
+  }
+}
