--- server/src/error/mysql-connection.error.ts
@@ -1,5 +1,0 @@
-export default class MysqlConnectionError extends Error {
-  constructor(msg: string) {
-    super(msg);
-  }
-}
