--- server/src/module/auth/application/dto/user.dto.ts
@@ -1,8 +1,0 @@
-export default class UserDTO {
-  constructor(
-    readonly id: string,
-    readonly names: string,
-    readonly surnames: string,
-    readonly email: string
-  ) {}
-}
