--- server/src/module/shared/domain/value-object/uuid.value-object.ts
@@ -1,18 +1,0 @@
-export default class UUID {
-  static #regex: RegExp = /[a-z0-9]{8}-[a-z0-9]{4}-[a-z0-9]{4}-[a-z0-9]{4}-[a-z0-9]{12}/;
-  readonly #value: string;
-
-  constructor(value: string) {
-    this.#value = value;
-    this.checkUUID();
-  }
-
-  private checkUUID(): void {
-    if (!UUID.#regex.test(this.#value))
-      throw new Error(`UUID Invalido: ${ this.#value }`);
-  }
-
-  getValue(): string {
-    return this.#value;
-  }
-}
